assignment1/requirements.txt
numpy
opencv-python
scipy
pillow
matplotlib